+++ packages/engine-formula/src/services/calculate-formula.service.ts
@@ -260,11 +260,11 @@ export class CalculateFormulaService extends Disposable {
             }
 
             const tree = treeList[i];
-            const astNode = tree.node;
+            const nodeData = tree.nodeData;
             const getDirtyData = tree.getDirtyData;
             let value: FunctionVariantType;
 
-            if (astNode == null && getDirtyData == null) {
+            if (nodeData == null && getDirtyData == null) {
                 throw new Error('AstNode or executor is null');
             }
 
@@ -288,11 +288,11 @@ export class CalculateFormulaService extends Disposable {
                 this._runtimeService.setRuntimeFeatureCellData(tree.featureId, runtimeCellData);
 
                 this._runtimeService.setRuntimeFeatureRange(tree.featureId, dirtyRanges);
-            } else if (astNode != null) {
-                if (interpreter.checkAsyncNode(astNode)) {
-                    value = await interpreter.executeAsync(astNode);
+            } else if (nodeData != null) {
+                if (interpreter.checkAsyncNode(nodeData.node)) {
+                    value = await interpreter.executeAsync(nodeData);
                 } else {
-                    value = interpreter.execute(astNode);
+                    value = interpreter.execute(nodeData);
                 }
 
                 if (tree.formulaId != null) {
@@ -326,6 +326,8 @@ export class CalculateFormulaService extends Disposable {
         } else if (!isArrayFormulaState) {
             this._runtimeService.markedAsNoFunctionsExecuted();
         }
+
+        this._runtimeService.clearReferenceAndNumberformatCache();
 
         return this._runtimeService.getAllRuntimeData();
     }