+++ packages/sheets-ui/src/controllers/render-controllers/sheet.render-controller.ts
@@ -15,8 +15,10 @@
  */
 
 import type { ICommandInfo, IExecutionOptions, IRange, Nullable, Workbook, Worksheet } from '@univerjs/core';
+import type { ISetFormulaCalculationNotificationMutation } from '@univerjs/engine-formula';
 import type { IAfterRender$Info, IBasicFrameInfo, IExtendFrameInfo, IRenderContext, IRenderModule, ISummaryFrameInfo, ITimeMetric, IViewportInfos, Scene } from '@univerjs/engine-render';
 import { CommandType, ICommandService, IContextService, Inject, Optional, Rectangle, RxDisposable } from '@univerjs/core';
+import { SetFormulaCalculationNotificationMutation } from '@univerjs/engine-formula';
 import {
     Rect,
     ScrollBar,
@@ -29,8 +31,8 @@ import {
 } from '@univerjs/engine-render';
 import { COMMAND_LISTENER_SKELETON_CHANGE, COMMAND_LISTENER_VALUE_CHANGE, MoveRangeMutation, SetRangeValuesMutation, SetWorksheetActiveOperation } from '@univerjs/sheets';
 import { ITelemetryService } from '@univerjs/telemetry';
-import { Subject, withLatestFrom } from 'rxjs';
 
+import { Subject, withLatestFrom } from 'rxjs';
 import {
     SHEET_COMPONENT_HEADER_LAYER_INDEX,
     SHEET_COMPONENT_MAIN_LAYER_INDEX,
@@ -423,6 +425,13 @@ export class SheetRenderController extends RxDisposable implements IRenderModule
 
     private _markUnitDirty(unitId: string, command: ICommandInfo) {
         const { mainComponent: spreadsheet, scene } = this._context;
+
+        if (command.id === SetFormulaCalculationNotificationMutation.id) {
+            const params = command.params as ISetFormulaCalculationNotificationMutation;
+            if (params.stageInfo != null) {
+                return;
+            }
+        }
 
         if (spreadsheet) {
             spreadsheet.makeDirty(); // refresh spreadsheet